+++ runtimes/server-interface/chat.ts
@@ -24,6 +24,7 @@ import {
     InlineChatResult,
     ContextCommandParams,
     CreatePromptParams,
+    InlineChatResultParams,
     ListConversationsParams,
     ListConversationsResult,
     ConversationClickParams,
@@ -32,6 +33,10 @@ import {
     GetSerializedChatParams,
     TabBarActionParams,
     TabBarActionResult,
+    ChatOptionsUpdateParams,
+    PromptInputOptionChangeParams,
+    ButtonClickParams,
+    ButtonClickResult,
 } from '../protocol'
 
 /**
@@ -46,6 +51,7 @@ export type Chat = {
     onEndChat: (handler: RequestHandler<EndChatParams, EndChatResult, void>) => void
     onQuickAction: (handler: RequestHandler<QuickActionParams, QuickActionResult, void>) => void
     openTab: (params: OpenTabParams) => Promise<OpenTabResult>
+    onButtonClick: (handler: RequestHandler<ButtonClickParams, ButtonClickResult, ButtonClickResult>) => void
     onListConversations: (handler: RequestHandler<ListConversationsParams, ListConversationsResult, void>) => void
     onConversationClick: (handler: RequestHandler<ConversationClickParams, ConversationClickResult, void>) => void
     onTabBarAction: (handler: RequestHandler<TabBarActionParams, TabBarActionResult, void>) => void
@@ -63,6 +69,9 @@ export type Chat = {
     onFollowUpClicked: (handler: NotificationHandler<FollowUpClickParams>) => void
     sendChatUpdate: (params: ChatUpdateParams) => void
     onFileClicked: (handler: NotificationHandler<FileClickParams>) => void
+    chatOptionsUpdate: (params: ChatOptionsUpdateParams) => void
     sendContextCommands: (params: ContextCommandParams) => void
     onCreatePrompt: (handler: NotificationHandler<CreatePromptParams>) => void
+    onInlineChatResult: (handler: NotificationHandler<InlineChatResultParams>) => void
+    onPromptInputOptionChange: (handler: NotificationHandler<PromptInputOptionChangeParams>) => void
 }